requirements.txt
numpy
Pillow
--find-links https://download.pytorch.org/whl/cpu/torch_stable.html
torch==1.7.1+cpu
torchvision==0.8.2+cpu
torch-summary
typing-extensions==3.7.4.3
ttach
tqdm
opencv-python
timm

torchinfo

# DETR
ipywidgets

# Yolo:
# Base ----------------------------------------
matplotlib>=3.2.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1

# Logging -------------------------------------
tensorboard>=2.4.1
# wandb

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Export --------------------------------------
# coremltools>=4.1  # CoreML export
# onnx>=1.9.0  # ONNX export
# onnx-simplifier>=0.3.6  # ONNX simplifier
# scikit-learn==0.19.2  # CoreML quantization
# tensorflow>=2.4.1  # TFLite export
# tensorflowjs>=3.9.0  # TF.js export

# Extras --------------------------------------
# albumentations>=1.0.3
# Cython  # for pycocotools https://github.com/cocodataset/cocoapi/issues/172
# pycocotools>=2.0  # COCO mAP
# roboflow
thop  # FLOPs computation
